feat: Remove state error problems Modify upload file process
